Create wallabag-sync.py
Should be all you need. 

The interesting thing, of course, is leaving this running on a server, and having it write your file to Dropbox (or the cloud service provider of your choice), which then syncs to your laptop, meaning everything you save in Wallabag gets indexed by Spotlight (or, again, the on-laptop search system of your choice).